Add New Aditions to build.sh
First, define te Motorolaa SKU target to be TARGET_PRODUCT as this is what is
used for compiling Motorola modules. Second, enable bulding with DDK and unsafe
headers as some modules require it. Third, to compile everything properly,
cange the location of where the moules in EXT_MODULES is sourced from (and
remove symlinks). Fourth, add the Motorola modules to the list of EXT_MODULES.
Finally, generate moto_product.bzl with mmi_product_name as this is required
for some Bazel files.